requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
ccxt==4.5.37
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
numpy==2.4.2
pandas==3.0.0
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
setuptools==82.0.0
six==1.17.0
ta==0.11.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
